Reestructure view_commands
