+++ src/pycture/commands/view_commands/__init__.py
@@ -1,0 +1,4 @@
+from .view_histogram import (ViewRedHistogram, ViewGreenHistogram, ViewBlueHistogram,
+                             ViewGrayScaleHistogram)
+from .view_image_info import (ViewImageBrightness, ViewImageSize, ViewImageContrast,
+                              ViewImageEntropy, ViewImageRanges)
